llm-api/requirements.txt
# LLM API Dependencies

# Core dependencies
langchain-core>=0.1.0
langchain-community>=0.0.10
pydantic>=2.0.0
requests>=2.25.0
colorama>=0.4.4

# LLM provider packages
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langchain-deepseek>=0.1.0
langchain-ollama>=0.1.0

# Optional dependencies
python-dotenv>=1.0.0
lmstudio>=0.1.0
streamlit>=1.28.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=1.0.0